No issue: mv ImageCarousel to separate file.
